requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.15
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
crcmod==1.7
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
decorator==4.4.2
Django==3.0.3
django-aliyun-oss2==0.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dlib==19.6.1
freeze==1.0.10
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
imageio==2.8.0
imutils==0.5.3
incremental==17.5.0
jmespath==0.9.5
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==3.2.0
msgpack==0.6.2
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.32
oss2==2.9.1
pandas==1.0.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
selenium==3.141.0
service-identity==18.1.0
signals==0.0.2
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
xlwt==1.3.0
zope.interface==4.7.2
